feat: send extracted CSV output to SFTP
